marketmind_backend/image-service/requirements.txt
Flask
Werkzeug
pyjwt
flask-cors
python-dotenv
pillow
numpy

